for incremental exp plot

+++ batch/check_pm25.py
@@ -4,19 +4,20 @@ import shutil
 import os 
 import subprocess
 
-### PASSED AGRUEMENTS
-if len(sys.argv) < 4:
-    print("you must set 4 arguments as stat[o3|rmse|...] [day1|day2|day3] start_date end_date event_equal[optional]")
+# PASSED N AGRUEMENTS IN : total number of argument including [0] is N+1
+if len(sys.argv) < 6:
+    print("you must set 5 arguments as stat[o3|rmse|...] [06|12|all] [day1|day2|day3|all] start_date end_date event_equal[optional]")
     sys.exit()
 else:
     stat_var = sys.argv[1]
-    flag_vday = sys.argv[2]
-    start_date = sys.argv[3]
-    end_date = sys.argv[4]
+    flag_cyc  = sys.argv[2]
+    flag_vday = sys.argv[3]
+    start_date = sys.argv[4]
+    end_date = sys.argv[5]
 if stat_var == "pm25":
     stat_var = "time_series"
-if len(sys.argv) > 5:
-    event_equal_flag = sys.argv[5]
+if len(sys.argv) > 6:
+    event_equal_flag = sys.argv[6]
 else:
     event_equal_flag = "true"
     print("the evn equalizer is set to true, if you want to change it please add true/false for the 5th argument")
@@ -32,6 +33,15 @@ fig_sdate = sdate.strftime(file_date_format)
 fig_edate = edate.strftime(file_date_format)
 METviewer_AWS_scripts_dir = "/lfs/h2/emc/vpppg/save/"+os.environ['USER']+"/METviewer_AWS"
 print(METviewer_AWS_scripts_dir)
+if flag_cyc == "all":
+   run_cycle=[ "06Z", "12Z" ]
+elif flag_cyc == "06":
+   run_cycle=[ "06Z" ]
+elif flag_cyc == "12":
+   run_cycle=[ "12Z" ]
+else:
+    print("verification verify day "+cyc+" not recongized.")
+    exit()
 if flag_vday == "all":
    vday=[ "day1", "day2" ]
 elif flag_vday == "day1":
@@ -44,34 +54,19 @@ else:
     print("verification verify day "+vday+" not recongized.")
     exit()
 
-vday=[ "day2" ]
-
 ## For diurnal cycle plot fix day as the # of fcst day
 if stat_var == "rmsedl":
    vday=[ "day3" ]
 elif stat_var == "medl":
    vday=[ "day3" ]
-
-run_cycle = [ "06Z", "12Z" ]
-run_cycle = [ "12Z" ]
+region = [ "CONUS" ]
 region = [ "FULL", "CONUS", "EAST", "WEST", "NEUS", "SEUS", "NWUS", "SWUS", "NEC", "SEC", "APL", "GMC", "LMV", "MDW", "NMT", "NPL", "SMT", "SPL", "NWC", "SWC", "SWD", "GRB" ]
 region = [ "Appalachia", "CONUS_Central", "CONUS_East", "CONUS", "CONUS_South", "CONUS_West", "CPlains", "DeepSouth", "GreatBasin", "GreatLakes", "Mezquital", "MidAtlantic", "NorthAtlantic", "NPlains", "NRockies", "PacificNW", "PacificSW", "Prairie", "Southeast", "Southwest", "SPlains", "SRockies" ]
-region = [ "CONUS" ]
-region = [ "CONUS_South" ]
-
-xml_data_dir = "/lfs/h2/emc/vpppg/save/"+os.environ['USER']+"/METviewer_AWS/my_plot_xmls_icmaq"
-xml_data_dir = "/lfs/h2/emc/vpppg/save/"+os.environ['USER']+"/METviewer_AWS/my_plot_xmls_v70c3"
-xml_data_dir = "/lfs/h2/emc/vpppg/save/"+os.environ['USER']+"/METviewer_AWS/my_plot_xmls_rrfs"
-xml_gen_python_name = "plot3.cmaq_pm25_"+stat_var.lower()+".py"
-xml_gen_python_name = "plot2.cmaq_pm25_"+stat_var.lower()+".py"
-xml_gen_python_name = "plot2.cmaq_b_pm25_"+stat_var.lower()+".py"
+region = [ "CONUS_Central", "CONUS_East", "CONUS", "CONUS_South", "CONUS_West" ]
+region = [ "NRockies", "SRockies", "GreatBasin", "PacificNW", "PacificSW", "CONUS_Central", "CONUS_East", "CONUS", "CONUS_South", "CONUS_West", "Southwest" ]
+xml_data_dir = "/lfs/h2/emc/vpppg/save/"+os.environ['USER']+"/METviewer_AWS/my_plot_xmls_base"
 plot_xml_file = "plot_cmaq_pm25_"+stat_var.lower()+".xml"
 scripts_dir = "/lfs/h2/emc/vpppg/save/"+os.environ['USER']+"/METviewer_AWS/script"
-
-checkfile=os.path.join(xml_data_dir,xml_gen_python_name)
-if not os.path.exists(checkfile):
-    print("Can not find "+checkfile)
-    sys.exit()
 
 batch_script_name = "mv_batch_on_aws.sh"
 tmp_data_dir = "/lfs/h2/emc/stmp/"+os.environ['USER']+"/run_batch_plot_"+stat_var.lower()
@@ -83,9 +78,40 @@ if os.path.exists(tmp_data_dir):
     shutil.rmtree(tmp_data_dir)
 os.makedirs(tmp_data_dir)
 
+database_header="mv_g2o_met_o3pm_b_"
+exp_labl=[ "LBL1", "LBL2" ]
+exp_symb=[ "MDL1", "MDL2" ]
+exp_incl=[ "prod", "v70c22" ]
+inum=len(exp_incl)
+xml_python_basename = "plot"+str(inum)+".cmaq_pm25_"+stat_var.lower()+".base"
+xml_gen_python_name = "common_plot.cmaq_pm25."+stat_var.lower()+".py"
+
+checkfile=os.path.join(xml_data_dir,xml_python_basename)
+if not os.path.exists(checkfile):
+    print("Can not find "+checkfile)
+    sys.exit()
+
 os.chdir(tmp_data_dir)
 shutil.copy(os.path.join(scripts_dir, batch_script_name), tmp_data_dir)
-shutil.copy(os.path.join(xml_data_dir, xml_gen_python_name), tmp_data_dir)
+shutil.copy(os.path.join(xml_data_dir, xml_python_basename), tmp_data_dir)
+
+iexp=0
+for i in exp_incl:
+    if iexp == 0 :
+        met_database_run=database_header+i.lower()+"_"+database_date
+    else:
+        met_database_run=met_database_run+','+database_header+i.lower()+"_"+database_date
+    iexp+=1
+run_command='sed -e "s!XYZdatabase_nameXYZ!'+met_database_run+'!" '
+for i in range(0,inum):
+    run_command=run_command+' -e "s!XYZ'+exp_symb[i]+'XYZ!'+exp_incl[i].upper()+'!" '
+    if i == 0 :
+        run_command=run_command+' -e "s!XYZ'+exp_labl[i]+'XYZ!'+exp_incl[i].upper()+'!" '
+    else:
+        run_command=run_command+' -e "s!XYZ'+exp_labl[i]+'XYZ!'+exp_incl[i].lower()+'!" '
+run_cmmmand=run_command+' '+xml_python_basename+' > '+xml_gen_python_name
+subprocess.call([run_cmmmand], shell=True)
+
 if 1 == 1:
     for area in region:
         for verf_day in vday:
@@ -104,5 +130,6 @@ if 1 == 1:
     else:
        ## partb=os.path.join("hchuang@rzdm:", "home", "www", "emc", "htdocs", "mmb", "hchuang", "transfer")
        partb=os.path.join("hchuang@rzdm:", "home", "www", "emc", "htdocs", "mmb", "hchuang", "ftp")
+       partb=os.path.join("hchuang@rzdm:", "home", "www", "emc", "htdocs", "mmb", "hchuang", "evs_verif", database_year, database_date )
 
     subprocess.call(['scp -p * '+partb], shell=True)
